# openapi.yaml
openapi: 3.0.3
info:
  title: backtests API
  description: >
    JSON HTTP API exposing the `backtest` and `info` CLI commands
    (see `internal/api/api.go`, started via `go run ./cmd serve -addr :8080`
    / `make serve`).


    This is the source of truth for the API contract used by the frontend.
    If anything here looks inconsistent with what the frontend was
    originally given, this document wins - treat any mismatch as a bug to
    raise, not something to route around.


    ## Conventions


    - All responses are `application/json`.
    - All monetary values (balances, profit, prices) are plain JSON numbers
      in BRL major units (e.g. `10000.00`, not cents, and not go-money's
      `{"amount":...,"currency":...}` object form).
    - Dates are `"YYYY-MM-DD"` strings throughout (requests and responses).
    - There is no authentication/authorization on any endpoint.
    - Errors, at any endpoint and for any failure, always have the shape
      documented in the `Error` schema below, with an appropriate HTTP
      status code:
      - `400` - invalid/missing input (bad JSON, missing required field,
        unparseable date/number, unknown strategy name, an invalid/missing
        `strategyParams` key for the chosen strategy, non-positive
        balance).
      - `404` - the request was well-formed but referred to something that
        doesn't exist (currently: an unimported/unknown asset on
        `GET /api/info`).
...
  version: "1.0.0"
servers:
...
paths:
  /api/info:
    get:
      operationId: getInfo
      summary: Look up the available imported date range for an asset.
      parameters:
        - name: asset
          in: query
          required: true
          description: >
            Ticker of the asset to look up, e.g. `PETR4`. Must match
            `^[A-Z0-9]+$` (uppercase letters and digits only) - this is
            enforced server-side to keep the value safe to use in a
            filesystem path lookup.
          schema:
            type: string
            pattern: "^[A-Z0-9]+$"
      responses:
        "200":
          description: The asset's imported date range.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InfoResponse"
              example:
                asset: PETR4
                earliest: "2010-01-04"
                latest: "2026-07-30"
        "400":
          description: >
            Missing/empty `asset` query param, or `asset` doesn't match
            `^[A-Z0-9]+$`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                missing:
                  value:
                    error: asset is required
                invalidFormat:
                  value:
                    error: 'asset must match ^[A-Z0-9]+$ (uppercase letters and digits only), got "../../etc"'
        "404":
          description: The asset has no imported data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: no data found for asset "NOPE9" (not imported yet?)
        "500":
          description: Unexpected server-side failure.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/assets:
    get:
      operationId: listAssets
      summary: >
        List available assets (by ticker), optionally filtered to a
        specific year and, when filtered, sorted by that year's trading
        volume.
      parameters:
        - name: year
          in: query
          required: false
          description: >
            When present, only tickers that have imported data for this
            year are returned. Omit to list every ticker with any imported
            data.
          schema:
            type: integer
          example: 2015
      responses:
        "200":
          description: >
            Tickers split into two groups: `stocks` (common equities, i.e.
            ordinary/preferred share classes) and `others` (units, ETFs,
            BDRs, index-tracking tickers, etc). Both arrays are `[]` (not
            `null`) when nothing matches - there is no `404` case for this
            endpoint. Sort order within each array depends on `year`: with
            no `year` filter, tickers are sorted alphabetically (there's no
            single year's volume to rank them by); with a `year` filter,
            tickers are sorted by that year's total trading volume,
            descending (most-traded first, ties broken alphabetically).


            Classification comes from `resources/cotahist/assets.json` (see
            `scripts/import_cotahist.go`, which bootstraps and enriches
            it from COTAHIST data on import): a ticker lands in `stocks`
            when its registry entry's `type` is `"stock"`, and in `others`
            otherwise - including when the ticker has no registry entry at
            all (e.g. it was imported before assets.json existed and the
            importer hasn't been re-run since).


            Each entry's `volume` (that year's total trading volume, in
            major currency units - same convention as a candle's `volume`
            field, see `Candle`) is only present when the request includes
            `year`; it's omitted entirely (not `null`) on an unfiltered
            request, since volume is meaningless without a specific year to
            sum over.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetsResponse"
              examples:
                noYearFilter:
                  summary: No `year` filter - volume omitted
                  value:
                    stocks: [{ ticker: "ITUB4" }, { ticker: "PETR3" }, { ticker: "PETR4" }, { ticker: "VALE3" }]
                    others: [{ ticker: "BOVA11" }, { ticker: "IBOV11" }]
                yearFilter:
                  summary: "`year=2015` - sorted by descending volume, volume present"
                  value:
                    stocks: [{ ticker: "PETR4", volume: 45123456.78 }, { ticker: "VALE3", volume: 30111222.33 }, { ticker: "ITUB4", volume: 12000000.0 }]
                    others: [{ ticker: "BOVA11", volume: 987654.32 }]
        "400":
          description: The `year` query param was present but not a valid integer.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: year must be a valid integer
        "500":
          description: Unexpected server-side failure.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/strategies:
    get:
      operationId: listStrategies
      summary: List the strategies accepted by `strategy` on `POST /api/backtest`.
      responses:
        "200":
          description: >
            A JSON array of strategy descriptors, sorted alphabetically by
            `name`. `params` describes the `strategyParams` keys that
            strategy expects (see `BacktestRequest.strategyParams` for the
            authoritative per-strategy prose) - `[]` for a strategy that
            takes none (e.g. `buy-and-hold`). This is a UI hint for
            rendering the right inputs, not the source of validation truth:
            the actual required-ness and cross-field rules (e.g.
            `shortPeriod` < `longPeriod`) are still enforced server-side by
            `POST /api/backtest` and surface as a normal `400` there.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StrategyInfo"
              example:
                - name: buy-and-hold
                  params: []
                - name: ema-crossover
                  params:
                    - key: shortPeriod
                      label: Short period
                      default: 12
                      min: 1
                      step: 1
                    - key: longPeriod
                      label: Long period
                      default: 26
                      min: 2
                      step: 1
                - name: ema-trend-breakout
                  params:
                    - key: shortPeriod
                      label: Short period
                      default: 8
                      min: 1
                      step: 1
                    - key: longPeriod
                      label: Long period
                      default: 80
                      min: 2
                      step: 1
                - name: rsi-threshold
                  params:
                    - key: period
                      label: RSI period
                      default: 14
                      min: 2
                      step: 1
                    - key: oversold
                      label: Oversold threshold
                      default: 30
                      min: 0
                      max: 100
                      step: 1
                    - key: overbought
                      label: Overbought threshold
                      default: 70
                      min: 0
                      max: 100
                      step: 1
                - name: sma-crossover
                  params:
                    - key: shortPeriod
                      label: Short period
                      default: 10
                      min: 1
                      step: 1
                    - key: longPeriod
                      label: Long period
                      default: 30
                      min: 2
                      step: 1
                - name: two-candle-breakout
                  params: []
        "500":
          description: Unexpected server-side failure.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/stop-losses:
    get:
      operationId: listStopLosses
      summary: List the type names accepted by `stopLoss.type` on `POST /api/backtest`.
      responses:
        "200":
          description: >
            A JSON array of stop-loss type name strings, sorted
            alphabetically.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
              example: ["fixed-amount", "none", "percent", "total-fixed-amount"]
        "500":
          description: Unexpected server-side failure.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/scan:
    post:
      operationId: runScan
      summary: >
        Run a strategy against every available asset (optionally restricted
        to a year), compared against a Buy & Hold baseline for each -
        equivalent to the `scan` CLI command. Same request shape as
        `POST /api/backtest` minus `asset` (a scan runs every asset, not
        one) plus `year`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanRequest"
            example:
              strategy: sma-crossover
              strategyParams: { shortPeriod: 10, longPeriod: 30 }
              stopLoss:
                type: percent
                value: 5
              start: "2015-01-01"
              end: "2015-12-30"
              balance: "10000.00"
              year: 2015
      responses:
        "200":
          description: >
            One entry per scanned asset (the asset universe from
            `cotahist.ListAssets(year)`, same as `GET /api/assets`), sorted
            with every successfully-scanned asset first (by `delta`
            descending - biggest winners first), then every asset whose
            backtest failed to load/run, last. This endpoint always returns
            `200` once the request itself is valid and the asset universe
            could be listed - an individual asset's backtest failing (e.g.
            corrupt imported data) does **not** fail the whole request, it
            just produces an error-only entry for that asset (see
            `ScanResult`). Deliberately lean: no per-operation `operations`
            arrays - drill into a specific asset via `POST /api/backtest` or
            the CLI's `compare` once a candidate is found here.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScanResult"
              example:
                - asset: PETR4
                  baselineProfitPercentage: -27.77
                  challengerProfitPercentage: 12.34
                  challengerTotalOperations: 6
                  delta: 40.11
                  won: true
                - asset: VALE3
                  baselineProfitPercentage: 8.1
                  challengerProfitPercentage: 8.1
                  challengerTotalOperations: 0
                  delta: 0
                  won: false
                - asset: XPTO3
                  error: "loading candles for XPTO3: ..."
        "400":
          description: >
            Invalid input - the same up-front validation
            `POST /api/backtest` performs (minus anything `asset`-specific,
            since there is no single `asset` here): missing required
            field(s) (`start`, `end`, `strategy`, `balance`), malformed JSON
            body, unparseable/non-positive `balance`, unknown `strategy`, a
            `strategyParams` missing a key `strategy` requires or with an
            out-of-range value, unparseable `start`/`end`, a `stopLoss`
            present without a `type`, an unknown `stopLoss.type`, or an
            invalid `stopLoss.value`. See `POST /api/backtest`'s `400`
            examples for the exact message shapes - they're produced by the
            same validation code.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                missingFields:
                  value:
                    error: "start, end, strategy and balance are all required"
                unknownStrategy:
                  value:
                    error: "unknown strategy \"nope\", available: buy-and-hold"
        "500":
          description: >
            Unexpected server-side failure listing the asset universe (e.g.
            `resources/cotahist` unreadable).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
    Error:
      type: object
      description: The JSON shape every endpoint returns on failure.
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message.
      example:
        error: human-readable message
    InfoResponse:
      type: object
      required: [asset, earliest, latest]
      properties:
        asset:
          type: string
          description: Echoes the request's `asset` query param.
          example: PETR4
        earliest:
          type: string
          format: date
          example: "2010-01-04"
        latest:
          type: string
          format: date
          example: "2026-07-30"
...
    AssetsResponse:
      type: object
      required: [stocks, others]
      properties:
        stocks:
          type: array
          description: >
            Common equities (ordinary/preferred share classes), e.g.
            `PETR4`, `PETR3`, `VALE3`, `ITUB4`.
          items:
            $ref: "#/components/schemas/AssetEntry"
          example: [{ ticker: "ITUB4" }, { ticker: "PETR3" }, { ticker: "PETR4" }, { ticker: "VALE3" }]
        others:
          type: array
          description: >
            Everything else - units, ETFs, BDRs, index-tracking tickers,
            e.g. `IBOV11`, `BOVA11`.
          items:
            $ref: "#/components/schemas/AssetEntry"
          example: [{ ticker: "BOVA11" }, { ticker: "IBOV11" }]
...